+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### [9.0.1] - unreleased
 
 ### Added
+- Multistep dialog feature, implemented on scanner dialog [#1725] (https://github.com/greenbone/gsa/pull/1725)
 
 ### Changed